rename SearchRecipes to CatalogFilters, SearchByTitle to SearchFilter, SearchByCategory to CategoriesFilter
