+++ src/App/components/CatalogFilters/CatalogFilters.module.scss
