+++ src/App/components/CatalogFilters/CategoriesFilter/CategoriesFilter.module.scss
