+++ src/App/components/CatalogFilters/CategoriesFilter/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CategoriesFilter';
