+++ src/App/components/CatalogFilters/SearchFilter/SearchFilter.module.scss
