+++ src/App/components/CatalogFilters/SearchFilter/SearchFilter.tsx
@@ -1,12 +1,12 @@
 import Input from "~App/components/Input";
 import Button from "~components/Button";
-import styles from './SearchByTitle.module.scss';
+import styles from './SearchFilter.module.scss';
 import { useCallback } from "react";
 import { observer } from "mobx-react-lite";
 import { useLocalStore } from "~utils/useLocalStore";
 import CatalogFiltersStore from "~store/CatalogStore/CatalogFiltersStore/CatalogFiltersStore";
 
-const SearchByTitle = () => {
+const SearchFilter = () => {
     const catalogFiltersStore = useLocalStore(() => new CatalogFiltersStore());
     const { tempSearch, setTempSearch, setSearch } = catalogFiltersStore;
 
@@ -32,4 +32,4 @@ const SearchByTitle = () => {
     )
 };
 
-export default observer(SearchByTitle);
+export default observer(SearchFilter);
