+++ src/App/components/CatalogFilters/SearchFilter/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SearchFilter';
