+++ src/App/components/CatalogFilters/types.ts
