--- src/App/components/SearchRecipes/SearchByFilter/index.ts
@@ -1,1 +1,0 @@
-export { default } from './SearchByFilter';
