--- src/App/components/SearchRecipes/SearchByTitle/index.ts
@@ -1,1 +1,0 @@
-export {default} from './SearchByTitle';
