--- src/App/components/SearchRecipes/index.ts
@@ -1,1 +1,0 @@
-export { default } from './SearchRecipes'
